% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_search.R
\name{param_search}
\alias{param_search}
\title{Request argument lists}
\usage{
param_search(param)
}
\arguments{
\item{param}{parameter names.}
}
\description{
Request argument lists
}
\examples{
# key <- 'Your key'
# param_search(param = 'state')

#' # param_search(param = 'year')

#' # param_search(param = 'category')
}
