% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramSearch.R
\name{paramSearch}
\alias{paramSearch}
\title{Return values of the selected input variables}
\usage{
paramSearch(param)
}
\arguments{
\item{param}{Name of input variable}
}
\value{
A data set that contains possible values of the selected input variable.
}
\description{
Return values of the selected input variables
}
\examples{
\dontrun{
key <- 'Your key' # Specify your API key here.
paramSearch(param = 'state')
paramSearch(param = 'year')
paramSearch(param = 'category')
}
}
